+++ src/ccl_power.c
@@ -538,6 +538,10 @@ void ccl_cosmology_compute_nonlin_power(ccl_cosmology* cosmo, ccl_f2d_t *psp_o,
         ccl_cosmology_compute_power_emu(cosmo, status);}
         break;
 
+      case ccl_pknl_from_input: {
+        ccl_cosmology_compute_nonlin_power_from_f2d(cosmo, psp_o, status);}
+        break;
+
     default: {
       *status = CCL_ERROR_INCONSISTENT;
       ccl_cosmology_set_status_message(
@@ -607,6 +611,12 @@ typedef struct {
   double R;
   int* status;
 } SigmaV_pars;
+
+// Params for k_NL integrand
+typedef struct {
+  ccl_cosmology *cosmo;
+  int* status;
+} KNL_pars;
 
 /* --------- ROUTINE: w_tophat ---------
 INPUT: kR, ususally a wavenumber multiplied by a smoothing radius
@@ -767,4 +777,61 @@ smoothed with a tophat filter of comoving size 8 Mpc/h
 
 double ccl_sigma8(ccl_cosmology *cosmo, int *status) {
   return ccl_sigmaR(cosmo, 8/cosmo->params.h, 1., status);
+}
+
+// Integrand for kNL integral
+static double kNL_integrand(double k,void *params) {
+  KNL_pars *par=(KNL_pars *)params;
+
+  double pk=ccl_linear_matter_power(par->cosmo,k, 1.,par->status);
+
+  return pk;
+}
+
+/* --------- ROUTINE: ccl_kNL ---------
+INPUT: cosmology, scale factor
+TASK: compute kNL, the scale for the non-linear cut
+*/
+double ccl_kNL(ccl_cosmology *cosmo,double a,int *status) {
+  if (!cosmo->computed_linear_power) {
+    *status = CCL_ERROR_LINEAR_POWER_INIT;
+    ccl_cosmology_set_status_message(
+      cosmo,
+      "ccl_power.c: ccl_sigmaR(): linear power spctrum has not been computed!");
+    return NAN;
+  }
+  if (!cosmo->computed_growth){
+    *status = CCL_ERROR_GROWTH_INIT;
+    ccl_cosmology_set_status_message(
+      cosmo,
+      "ccl_power.c: ccl_sigmaR(): growth factor splines have not been precomputed!");
+    return NAN;
+  }
+
+  KNL_pars par;
+  par.status = status;
+
+  par.cosmo=cosmo;
+  gsl_integration_cquad_workspace *workspace =  NULL;
+  gsl_function F;
+  F.function=&kNL_integrand;
+  F.params=&par;
+  double PL_integral;
+
+  workspace = gsl_integration_cquad_workspace_alloc(cosmo->gsl_params.N_ITERATION);
+  if (workspace == NULL) {
+    *status = CCL_ERROR_MEMORY;
+  }
+  if (*status == 0) {
+    int gslstatus = gsl_integration_cquad(&F, cosmo->spline_params.K_MIN, cosmo->spline_params.K_MAX,
+                                          0.0, cosmo->gsl_params.INTEGRATION_KNL_EPSREL,
+                                          workspace,&PL_integral,NULL,NULL);
+    if(gslstatus != GSL_SUCCESS) {
+      ccl_raise_gsl_warning(gslstatus, "ccl_power.c: ccl_kNL():");
+      *status |= gslstatus;
+    }
+  }
+  gsl_integration_cquad_workspace_free(workspace);
+  double sigma_eta = sqrt(PL_integral/(6*M_PI*M_PI))*ccl_growth_factor(cosmo, a, status);
+  return pow(sigma_eta, -1);
 }